pyproject.toml
[project]
name = "nanogate"
version = "0.1.0"
description = "Multi-agent API gateway over hkuds/nanobot"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "nanobot contributors"}
]
keywords = ["ai", "agent", "gateway", "nanobot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "nanobot-ai @ git+https://github.com/hkuds/nanobot.git",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.12.0",
    "httpx>=0.28.0",
    "docker>=7.1.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-xdist>=3.5",
    "aiohttp>=3.9",
]

[project.scripts]
nanogate = "gateway.server:main"
nanogate-agent = "agent.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["gateway", "agent"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests that require a running nanogate gateway (deselect with -m 'not integration')",
]
